admin: Extract a shared ContentScopeDataGrid component
The permission-specific content scopes dialog had its own DataGrid setup
that drifted from the assigned scopes grid. Extract a shared
ContentScopeDataGrid that owns the content scope dimension columns,
pagination and an optional checkbox selection, and render it from both the
assigned scopes grid and the override content scopes dialog so they stay in
sync.
