+++ packages/admin/cms-admin/src/userPermissions/user/permissions/ContentScopeDataGrid.tsx
@@ -1,0 +1,161 @@
+import { DataGridToolbar, FillSpace, type GridColDef, GridToolbarQuickFilter } from "@comet/admin";
+import { Typography } from "@mui/material";
+import type { GridRowSelectionModel, GridToolbarProps } from "@mui/x-data-grid";
+import type { ReactNode } from "react";
+import { FormattedMessage } from "react-intl";
+
+import { DataGrid } from "../../../dataGrid/DataGrid";
+import { camelCaseToHumanReadable } from "../../utils/camelCaseToHumanReadable";
+import type { GQLAvailableContentScopesQuery } from "./selectScopesDialogContent/SelectScopesDialogContent.generated";
+
+export type ContentScope = {
+    [key: string]: string;
+};
+
+/**
+ * Wildcard value a content scope dimension can have when `getContentScopesForUser` grants access to any value for it.
+ * Must match `UserPermissions.allValues` in `@comet/cms-api`.
+ */
+const contentScopeAllValues = "*";
+
+interface ToolbarProps extends GridToolbarProps {
+    toolbarAction?: ReactNode;
+}
+
+function ContentScopeDataGridToolbar({ toolbarAction }: ToolbarProps) {
+    return (
+        <DataGridToolbar>
+            <GridToolbarQuickFilter />
+            <FillSpace />
+            {toolbarAction}
+        </DataGridToolbar>
+    );
+}
+
+interface ContentScopeDataGridSelection {
+    selectedRowIds: string[];
+    onSelectedRowIdsChange: (selectedRowIds: string[]) => void;
+    disabled?: boolean;
+}
+
+interface ContentScopeDataGridProps {
+    rows: ContentScope[];
+    availableContentScopes: GQLAvailableContentScopesQuery["availableContentScopes"];
+    availableContentScopeDimensions?: Array<{ name: string; label: string }>;
+    hasAllContentScopes?: boolean;
+    /** Additional columns appended after the content scope dimension columns (e.g. assignment type or actions). */
+    additionalColumns?: GridColDef<ContentScope>[];
+    /** Rendered on the right of the toolbar (e.g. an "Add scope" button). */
+    toolbarAction?: ReactNode;
+    /** When set, the grid shows a checkbox for each row and reports the selected rows by their id. */
+    selection?: ContentScopeDataGridSelection;
+}
+
+/**
+ * Renders content scopes in a consistent grid: one column per content scope dimension, pagination and an optional
+ * checkbox selection. Shared by the assigned scopes grid and the permission-specific content scopes dialog.
+ */
+export function ContentScopeDataGrid({
+    rows,
+    availableContentScopes,
+    availableContentScopeDimensions,
+    hasAllContentScopes,
+    additionalColumns = [],
+    toolbarAction,
+    selection,
+}: ContentScopeDataGridProps) {
+    const columns: GridColDef<ContentScope>[] = [
+        ...generateGridColumnsFromContentScopeProperties(availableContentScopes, {
+            contentScopes: rows,
+            dimensions: availableContentScopeDimensions,
+            hasAllContentScopes,
+        }),
+        ...additionalColumns,
+    ];
+
+    return (
+        <DataGrid<ContentScope>
+            rows={rows}
+            columns={columns}
+            loading={false}
+            getRowId={(row) => JSON.stringify(row)}
+            pagination
+            pageSizeOptions={[10, 25, 50]}
+            initialState={{ pagination: { paginationModel: { pageSize: 10 } } }}
+            slots={{ toolbar: ContentScopeDataGridToolbar }}
+            slotProps={{ toolbar: { toolbarAction } as ToolbarProps }}
+            showToolbar
+            checkboxSelection={selection ? !selection.disabled : undefined}
+            rowSelectionModel={selection ? { type: "include", ids: new Set<string>(selection.selectedRowIds) } : undefined}
+            onRowSelectionModelChange={
+                selection
+                    ? (selectionModel: GridRowSelectionModel) =>
+                          selection.onSelectedRowIdsChange(Array.from(selectionModel.ids).map((id) => String(id)))
+                    : undefined
+            }
+            disableRowSelectionExcludeModel={selection !== undefined}
+        />
+    );
+}
+
+export function generateGridColumnsFromContentScopeProperties(
+    availableContentScopes: GQLAvailableContentScopesQuery["availableContentScopes"],
+    {
+        contentScopes = [],
+        dimensions = [],
+        hasAllContentScopes = false,
+    }: {
+        contentScopes?: ContentScope[];
+        dimensions?: Array<{ name: string; label: string }>;
+        hasAllContentScopes?: boolean;
+    } = {},
+): GridColDef[] {
+    const dimensionLabelsByName = new Map(dimensions.map((dimension) => [dimension.name, dimension.label]));
+    // Declared dimensions are the primary source so that dimensions not present in any value (e.g. an optional wildcard-only
+    // dimension) still get a column; keys of the available and displayed scopes are unioned in as a fallback.
+    const uniquePropertyNames = Array.from(
+        new Set([
+            ...dimensions.map((dimension) => dimension.name),
+            ...availableContentScopes.flatMap((item) => Object.keys(item.scope)),
+            ...contentScopes.flatMap((scope) => Object.keys(scope)),
+        ]),
+    );
+    return uniquePropertyNames.map((propertyName) => {
+        return {
+            field: propertyName,
+            flex: 1,
+            pinnable: false,
+            sortable: false,
+            filterable: true,
+            headerName: dimensionLabelsByName.get(propertyName) ?? camelCaseToHumanReadable(propertyName),
+            renderCell: ({ row }: { row: ContentScope }) => {
+                const value = row[propertyName];
+                const isAllValues =
+                    value === contentScopeAllValues ||
+                    // A user with all content scopes also has all values for dimensions that are not part of the available
+                    // content scopes (e.g. an optional dimension), which are therefore not set on the scope.
+                    (value === undefined && hasAllContentScopes);
+                if (isAllValues) {
+                    return (
+                        <Typography variant="body2">
+                            <FormattedMessage id="comet.userPermissions.allContentScopeValues" defaultMessage="All" />
+                        </Typography>
+                    );
+                }
+
+                // A wildcard dimension prevents the whole scope from matching an available scope, so labels are resolved per dimension
+                const label = availableContentScopes.find((availableContentScope) => availableContentScope.scope[propertyName] === value)?.label[
+                    propertyName
+                ];
+                if (label) {
+                    return <Typography variant="body2">{label}</Typography>;
+                }
+                // A value without a label is a free value of a dimension that is not part of the available content scopes
+                if (value !== undefined) {
+                    return <Typography variant="body2">{String(value)}</Typography>;
+                }
+                return "-";
+            },
+        };
+    });
+}
